+++ .gitignore
@@ -1,3 +1,4 @@
 env
 *.pyc
 data/
+.idea
