+++ lm_eval/tasks/race.py
@@ -1,7 +1,8 @@
 import collections
 import datasets
 import numpy as np
-from lm_eval.base import rf, mean
+from lm_eval.base import rf
+from ..metrics import mean
 from . common import HFTask
 
 import os
