+++ lm_eval/tasks/triviaqa.py
@@ -1,7 +1,8 @@
 import os
 import json
 import random
-from lm_eval.base import Task, mean, rf
+from lm_eval.base import Task, rf
+from ..metrics import mean
 from ..utils import sh
 
 class TriviaQA(Task):
