+++ lm_eval/tasks/wsc273.py
@@ -1,6 +1,7 @@
 import numpy as np
 import random
-from lm_eval.base import rf, mean
+from lm_eval.base import rf
+from ..metrics import mean
 from . common import HFTask
 
 """
